# Docs/API_Endpoints/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ProductLinkRequest:
      type: object
      properties:
        mainProductId:
          $ref: "#/components/schemas/UUID"
        subProductId:
          $ref: "#/components/schemas/UUID"
...
    ProductRequestSimple:
      type: object
      properties:
        price:
          type: number
          format: double
        displayName:
          type: string
        symbolIdentifier:
          type: string
        minimalWarning:
          type: integer
          format: int32
    ProductResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        price:
          type: number
          format: double
        displayName:
          type: string
        symbolIdentifier:
          type: string
        minimalWarning:
          type: integer
          format: int32
        productCount:
          type: integer
          format: int32
        subProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /product:
    post:
      summary: Create a Product for the currently logged-in Food Court Worker Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequestSimple"
        required: true
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  price:
                    type: number
                    format: double
                  displayName:
                    type: string
                  symbolIdentifier:
                    type: string
                  minimalWarning:
                    type: integer
                    format: int32
                  productCount:
                    type: integer
                    format: int32
                  subProducts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
...
  /product/by_id/{id}:
    delete:
      summary: Delete a Product by its ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Product deleted
          content:
            text/plain: {}
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
        - ADMIN
...
  /product/{id}:
    get:
      summary: Get a Product by ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  price:
                    type: number
                    format: double
                  displayName:
                    type: string
                  symbolIdentifier:
                    type: string
                  minimalWarning:
                    type: integer
                    format: int32
                  productCount:
                    type: integer
                    format: int32
                  subProducts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - GUEST
  /products/assignments:
    delete:
      summary: Delete an assignment by Main/Sub-ProductIds
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductLinkRequest"
        required: true
      responses:
        "200":
          description: Assignment deleted
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "404":
          description: Assignment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Assignment Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
        - ADMIN
    post:
      summary: Create a Main/Sub-Product Assignment (by IDs)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductLinkRequest"
        required: true
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                type: boolean
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Assignment Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
        - ADMIN
  /products/assignments/by_id/{id}:
    delete:
      summary: Delete an Assignment by its ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Assignment deleted
          content:
            text/plain: {}
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "404":
          description: Assignment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Assignment Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
        - ADMIN
...
info:
  title: App API
  version: 0.1.0-SNAPSHOT